feat(hr): Fix login and adding logic middleware auth

+++ core/routes/api.php
@@ -37,9 +37,9 @@ Route::prefix('v1')->group(function () {
 
     Route::controller(KaryawanController::class)->group(function () {
         Route::prefix('employee')->group(function () {
-            Route::post('store', 'store')->middleware(['auth:sanctum', 'ability:karyawan:outsource']);
-            Route::put('{id}/update', 'update')->middleware(['auth:sanctum']);
-            Route::delete('{id}/delete', 'destroy')->middleware(['auth:sanctum']);
+            Route::post('store', 'store')->middleware(['auth:sanctum', 'ability:user_type:admin,user_type:hr']);
+            Route::put('{id}/update', 'update')->middleware(['auth:sanctum', 'ability:user_type:admin,user_type:hr']);
+            Route::delete('{id}/delete', 'destroy')->middleware(['auth:sanctum', 'ability:user_type:admin,user_type:hr']);
         });
     });
 });
